debug

+++ miniPr/MODEL/prmodel.h
@@ -30,8 +30,7 @@ public:
 	void changespeedVideo(double rate);
 	void cutVideo(int start_frame, int end_frame);
 	void integrateAllVideo();
-	void addSubtitle(const string &text, int start_frame, int end_frame, int pos_x, int pos_y, int color_r, int color_g, int color_b, int size, bool isbold);
-
+	void addSubtitle(const string &text, int start_frame, int end_frame, double pos_x, double pos_y, double color_r, double color_g, double color_b, int size, bool isbold);
 	void setEditIndex(int index);
